update - moving table components to directory for organization purposes
