+++ frontend/components/census_tables/table6.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 
-let JsonData= [
+let JsonData = [
     {
         "subject": "VALUE OF HOME"
     },
@@ -358,10 +358,10 @@ let JsonData= [
     }
 ];
 
-function Table6(){
-    const DisplayData=JsonData.map(
-        (info)=>{
-            return(
+function Table6() {
+    const DisplayData = JsonData.map(
+        (info) => {
+            return (
                 <tr key={info.count}>
                     <td>{info.subject}</td>
                     <td>{info.tract78}</td>
@@ -370,19 +370,19 @@ function Table6(){
         }
     );
 
-    return(
+    return (
         <div>
             <table className="table formatting">
                 <thead>
-                    <tr>
-                        <th>Subject</th>
-                        <th>Tract 78</th>
-                    </tr>
+                <tr>
+                    <th>Subject</th>
+                    <th>Tract 78</th>
+                </tr>
                 </thead>
                 <tbody>
 
 
-                    {DisplayData}
+                {DisplayData}
 
                 </tbody>
             </table>
@@ -390,4 +390,5 @@ function Table6(){
         </div>
     );
 }
+
 export default Table6;
